refactor([DST-1419]): pagination in its own Panel.Content for padding
The bleed Panel.Content correctly removes horizontal padding for the
edge-to-edge table, but the pagination inherited that bleed and lost
its own padding. Split the pagination into a separate component so
page.tsx can render it in a non-bleed Panel.Content.

- Extract the search/filter/sort/page pipeline into hooks/useVenues.ts
  (matches the production template's useOrganizers boundary). Safe to
  call from multiple components — nuqs dedups URL subscriptions.
- New VenuesPagination component consumes useVenues for totalItems +
  usePagination for setters; owns the 3-column grid footer.
- VenuesTable now renders just the table inside a region wrapper. Uses
  a stable id (VENUES_REGION_ID) so VenuesPagination can scroll back
  to it on page change without a passed-down ref.
- page.tsx adds a third Panel.Content (non-bleed) for the pagination.
